requirements.txt
# Python
requests==2.31.0
boto3==1.34.0
snowflake-connector-python==3.6.0
python-dotenv==1.0.0

# dbt
dbt-snowflake==1.7.0

# Streamlit
streamlit==1.31.0
pandas==2.1.0

# Airflow
# apache-airflow==2.8.1
